requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0
supabase>=2.26.0
requests==2.31.0
polar-sdk>=0.28.0
boto3==1.34.14
schedule==1.2.0
python-dateutil==2.8.2
